requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.7.0
absl-py=2.3.1=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2025.7.15=hca03da5_0
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
expat=2.7.1=h313beb8_0
filelock=3.19.1=pypi_0
fonttools=4.59.2=pypi_0
fsspec=2025.7.0=pypi_0
grpcio=1.74.0=pypi_0
imageio=2.34.1=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.5.2=pypi_0
kiwisolver=1.4.9=pypi_0
lazy-loader=0.4=pypi_0
libcxx=20.1.8=h8869778_0
libffi=3.4.4=hca03da5_1
markdown=3.8.2=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.8.4=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.5=hee39554_0
networkx=3.4.2=pypi_0
numpy=1.26.4=pypi_0
opencv-python-headless=4.9.0.80=pypi_0
openssl=3.0.17=h4ee41c1_0
packaging=25.0=pypi_0
pandas=2.1.4=pypi_0
pillow=10.3.0=pypi_0
pip=25.2=pyhc872135_0
protobuf=6.32.0=pypi_0
pyparsing=3.2.3=pypi_0
python=3.10.18=h19e8193_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2025.2=pypi_0
readline=8.3=h0b18652_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.15.3=pypi_0
setuptools=80.9.0=pypi_0
six=1.17.0=pypi_0
sqlite=3.50.2=h79febb2_1
sympy=1.14.0=pypi_0
tensorboard=2.20.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
threadpoolctl=3.6.0=pypi_0
tifffile=2025.5.10=pypi_0
tk=8.6.15=hcd8a7d5_0
torch=2.2.2=pypi_0
torchvision=0.17.2=pypi_0
tqdm=4.66.4=pypi_0
typing-extensions=4.15.0=pypi_0
tzdata=2025.2=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py310hca03da5_0
xz=5.6.4=h80987f9_1
zlib=1.2.13=h18a0788_1
